# lightshield/web/routes.py
"""LightShield Web API — 路由蓝图。

所有 API 端点注册在 /api 前缀下：
- POST /api/login          — 用户登录（公开）
- POST /api/logout         — 用户登出（公开）
- POST /api/scan           — 提交扫描任务（需登录）
- GET  /api/scan/<task_id> — 查询任务状态（需登录）
- GET  /api/report/<scan_id> — 获取扫描报告（需登录）

响应格式统一为 JSON 信封：
  成功: {"success": true, ...}
  失败: {"error": true, "message": "...", "code": N}
"""

from __future__ import annotations

import json
import secrets
import time
from datetime import datetime, timezone
from fnmatch import fnmatchcase
from pathlib import Path

from flask import Blueprint, Response, current_app, jsonify, request, send_from_directory, session, stream_with_context

from lightshield.config import LightShieldConfig
from lightshield.core import R4_CONFIRMATION_REQUIRED
from lightshield.report.reporter import ReportGenerator
from lightshield.utils.constants import ScanStatus
from lightshield.utils.logger import get_logger
from lightshield.utils.validator import TargetValidator
from lightshield.web.auth import login, login_required, logout
from lightshield.web.csrf import csrf_exempt, csrf_protect

# ---------------------------------------------------------------------------
# 蓝图
# ---------------------------------------------------------------------------

api_bp = Blueprint("api", __name__, url_prefix="/api")

SCRIPT_FILENAME_PATTERNS = ("harden_*.sh", "harden_*.ps1", "rollback_*.sh", "rollback_*.ps1")
SSE_FINAL_STATUSES = {ScanStatus.COMPLETED.value, ScanStatus.PARTIAL.value, ScanStatus.FAILED.value}
SSE_INTERVAL_SECONDS = 0.5
SSE_TIMEOUT_SECONDS = 20.0


# =============================================================================
# 鉴权端点（公开）
# =============================================================================


@api_bp.route("/login", methods=["POST"])
@csrf_exempt
def api_login():
    """用户登录。

    Request JSON:
        {"username": "admin", "password": "lightshield"}

    Response 200:
        {"success": true, "message": "登录成功"}
    Response 400:
        {"error": true, "message": "请提供用户名和密码", "code": 400}
    Response 401:
        {"error": true, "message": "用户名或密码错误", "code": 401}
    """
    data = request.get_json(silent=True)
    if not data:
        return jsonify({"error": True, "message": "请求体不能为空，需要 JSON", "code": 400}), 400

    username = data.get("username", "")
    password = data.get("password", "")

    # C-003: 类型校验——防止 null/数字等非字符串凭证传入 login() 导致 TypeError
    if not isinstance(username, str) or not isinstance(password, str):
        return jsonify({"error": True, "message": "用户名和密码必须为字符串", "code": 400}), 400

    if not username or not password:
        return jsonify({"error": True, "message": "请提供用户名和密码", "code": 400}), 400

    if login(username, password):
        return jsonify({"success": True, "message": "登录成功"}), 200

    return jsonify({"error": True, "message": "用户名或密码错误", "code": 401}), 401


@api_bp.route("/logout", methods=["POST"])
@csrf_exempt
def api_logout():
    """用户登出。

    Response 200:
        {"success": true, "message": "已登出"}
    """
    logout()
    return jsonify({"success": True, "message": "已登出"}), 200


# =============================================================================
# 扫描端点（需登录）
# =============================================================================


@api_bp.route("/scan", methods=["POST"])
@login_required
def api_submit_scan():
    """提交扫描任务。

    Request JSON:
        {
            "target": "192.168.1.1",          // 必填：目标 IP/域名
            "scan_types": ["port_scan", ...],  // 可选：扫描类型列表，默认全部
            "confirm_ownership": true          // 必填：已获得目标所有权或明确授权
        }

    Response 202:
        {"task_id": "LS-20260614-...", "status": "accepted", "target": "..."}
    Response 400:
        {"error": true, "message": "...", "code": 400}
    """
    data = request.get_json(silent=True)
    if not data:
        return jsonify({"error": True, "message": "请求体不能为空，需要 JSON", "code": 400}), 400

    target = (data.get("target") or "").strip()
    if not target:
        return jsonify({"error": True, "message": "请提供扫描目标 (target)", "code": 400}), 400

    # R2 合规：双层校验（API 层 + core.submit_scan 内部再校验）
    is_valid, reason = TargetValidator.validate(target)
    if not is_valid:
        return jsonify({"error": True, "message": f"[R2 违规] {reason}", "code": 400}), 400

    # C-004: scan_types 类型校验——防止传入字符串/整数等非列表类型
    raw_scan_types = data.get("scan_types")
    if raw_scan_types is not None and not isinstance(raw_scan_types, list):
        return jsonify(
            {
                "error": True,
                "message": f"scan_types 必须是字符串数组，收到类型: {type(raw_scan_types).__name__}",
                "code": 400,
            }
        ), 400
    # C-004: 列表内每个元素也须为字符串
    if isinstance(raw_scan_types, list):
        for i, item in enumerate(raw_scan_types):
            if not isinstance(item, str):
                return jsonify(
                    {
                        "error": True,
                        "message": f"scan_types[{i}] 必须是字符串，收到类型: {type(item).__name__}",
                        "code": 400,
                    }
                ), 400

    if not _is_truthy(data.get("confirm_ownership")):
        return jsonify({"error": True, "message": R4_CONFIRMATION_REQUIRED, "code": 400}), 400

    scan_types: list[str] | None = raw_scan_types

    core = current_app.config["LIGHTSHIELD_CORE"]
    try:
        task_id = core.submit_scan(
            target=target,
            scan_types=scan_types,
            confirm_ownership=True,
        )
    except Exception as exc:
        return jsonify({"error": True, "message": f"扫描提交失败：{exc}", "code": 500}), 500

    return jsonify({"task_id": task_id, "status": "accepted", "target": target}), 202


@api_bp.route("/scan/<task_id>", methods=["GET"])
@login_required
def api_get_scan_status(task_id: str):
    """查询扫描任务状态。

    Response 200:
        {"task_id": "...", "status": "completed", "target": "...", ...}
    Response 404:
        {"error": true, "message": "任务不存在: LS-xxx", "code": 404}
    """
    core = current_app.config["LIGHTSHIELD_CORE"]
    status = core.get_scan_status(task_id)

    if status.get("status") == "not_found":
        return jsonify({"error": True, "message": f"任务不存在: {task_id}", "code": 404}), 404

    return jsonify(status), 200


@api_bp.route("/scan/<task_id>/stream", methods=["GET"])
@login_required
def api_scan_stream(task_id: str):
    """SSE 实时推送扫描进度。"""
    core = current_app.config["LIGHTSHIELD_CORE"]
    logger = get_logger()

    def _events():
        started_at = time.monotonic()
        logger.info("web", f"SSE 扫描进度连接已建立：task_id={task_id}")

        while True:
            try:
                status = core.get_scan_status(task_id)
            except Exception as exc:
                logger.error("web", f"SSE 扫描状态读取失败：task_id={task_id}", exception=exc)
                yield _format_sse({"message": "扫描状态读取失败", "code": 500}, event="error")
                return

            state = str(status.get("status", "unknown"))
            if state == "not_found":
                logger.warning("web", f"SSE 任务不存在：task_id={task_id}")
                yield _format_sse({"message": f"任务不存在: {task_id}", "code": 404}, event="error")
                return

            yield _format_sse(status)

            if state in SSE_FINAL_STATUSES:
                logger.info("web", f"SSE 扫描进度完成：task_id={task_id} status={state}")
                yield _format_sse(status, event="done")
                return

            if time.monotonic() - started_at >= SSE_TIMEOUT_SECONDS:
                logger.warning("web", f"SSE 扫描进度超时：task_id={task_id} status={state}")
                yield _format_sse(
                    {
                        "task_id": task_id,
                        "status": state,
                        "message": "扫描进度监听超时，请稍后刷新或查看历史记录",
                    },
                    event="timeout",
                )
                return

            time.sleep(SSE_INTERVAL_SECONDS)

    return Response(
        stream_with_context(_events()),
        mimetype="text/event-stream",
        headers={
            "Cache-Control": "no-cache",
            "X-Accel-Buffering": "no",
        },
    )


# =============================================================================
# 报告端点（需登录）
# =============================================================================


@api_bp.route("/report/<scan_id>", methods=["GET"])
@login_required
def api_get_report(scan_id: str):
    """获取扫描报告。

    Query params:
        format: "markdown"（默认）、"text" 或 "pdf"

    Response 200:
        报告内容 (Markdown/Text: text/plain; PDF: application/pdf)
    Response 404:
        {"error": true, "message": "扫描记录不存在: LS-xxx", "code": 404}
    Response 409:
        {"error": true, "message": "扫描尚未完成，无法生成报告。当前状态: failed", "code": 409}
    """
    config: LightShieldConfig = current_app.config["LIGHTSHIELD_CONFIG"]
    core = current_app.config["LIGHTSHIELD_CORE"]

    scan = core.load_scan(scan_id)
    if scan is None:
        return jsonify({"error": True, "message": f"扫描记录不存在: {scan_id}", "code": 404}), 404

    # 仅允许已完成或部分完成的扫描生成报告
    if scan.status not in (ScanStatus.COMPLETED, ScanStatus.PARTIAL):
        return jsonify(
            {
                "error": True,
                "message": f"扫描尚未完成，无法生成报告。当前状态: {scan.status.value}",
                "code": 409,
            }
        ), 409

    # 生成报告
    fmt = request.args.get("format", "markdown")
    if fmt not in ("markdown", "text", "pdf"):
        fmt = "markdown"

    reporter = ReportGenerator(output_dir=config.report_output_dir)
    try:
        report = reporter.generate(scan, findings=scan.findings, fmt=fmt)
    except Exception as exc:
        return jsonify({"error": True, "message": f"报告生成失败：{exc}", "code": 500}), 500

    if fmt == "pdf":
        return (
            report,
            200,
            {
                "Content-Type": "application/pdf",
                "Content-Disposition": f'attachment; filename="lightshield-{scan_id}.pdf"',
            },
        )

    return report, 200, {"Content-Type": "text/plain; charset=utf-8"}


@api_bp.route("/harden/<scan_id>", methods=["POST"])
@login_required
def api_generate_harden(scan_id: str):
    """生成加固和回滚脚本，返回脚本路径。"""
    data = request.get_json(silent=True) or request.form.to_dict() or {}
    os_platform = str(data.get("os_platform", "linux")).lower()
    if os_platform not in ("linux", "windows"):
        return jsonify({"error": True, "message": "不支持的操作系统，请选择 linux 或 windows", "code": 400}), 400

    if not _is_truthy(data.get("confirm_ownership")):
        return jsonify({"error": True, "message": "[R4] 请先确认目标所有权或授权范围", "code": 400}), 400

    core = current_app.config["LIGHTSHIELD_CORE"]
    scan = core.load_scan(scan_id)
    if scan is None:
        return jsonify({"error": True, "message": f"扫描记录不存在: {scan_id}", "code": 404}), 404

    recommendations = core.get_recommendations(scan_id)
    if not recommendations:
        return jsonify({"success": True, "generated": False, "message": "未发现需要加固的风险项", "code": 200}), 200

    config: LightShieldConfig = current_app.config["LIGHTSHIELD_CONFIG"]
    session["harden_confirmed_at"] = datetime.now(timezone.utc).isoformat()

    try:
        result = core.generate_hardening(
            scan.target,
            findings=scan.findings,
            recommendations=recommendations,
            output_dir=config.report_output_dir,
            os_platform=os_platform,
        )
    except Exception as exc:
        return jsonify({"error": True, "message": f"脚本生成失败：{exc}", "code": 500}), 500

    status = result.status.value if hasattr(result.status, "value") else str(result.status)
    script_filename = _script_basename(result.script_path)
    rollback_filename = _script_basename(result.rollback_path)
    return jsonify(
        {
            "success": True,
            "generated": True,
            "action_count": result.action_count,
            "script_path": result.script_path,
            "rollback_path": result.rollback_path,
            "script_filename": script_filename,
            "rollback_filename": rollback_filename,
            "status": status,
            "message": f"已生成 {result.action_count} 条加固操作",
        }
    ), 200


@api_bp.route("/harden/<scan_id>/verify", methods=["POST"])
@login_required
@csrf_protect
def api_harden_closed_loop(scan_id: str):
    """v0.0.40 加固闭环——触发 run_harden_closed_loop，返回全链路结果。

    请求体（JSON）：
        mode: "dry_run" | "apply"（默认 dry_run）
        os_platform: "linux" | "windows"（默认 linux）
        confirm_ownership: bool（APPLY 必须 true）
        confirm_execute: bool（APPLY 必须 true）

    APPLY 模式缺双确认 → 400 + 错误文案。
    DRY_RUN 模式 → overall="generated_only"，不改系统。
    """
    from lightshield.utils.constants import OSPlatform

    logger = get_logger()
    data = request.get_json(silent=True) or request.form.to_dict() or {}

    # ---- 参数解析 + 校验 ----
    mode = str(data.get("mode", "dry_run")).lower()
    if mode not in ("dry_run", "apply"):
        return jsonify({"error": True, "message": "mode 必须是 dry_run 或 apply", "code": 400}), 400

    os_platform_str = str(data.get("os_platform", "linux")).lower()
    try:
        os_platform = OSPlatform(os_platform_str)
    except ValueError:
        return jsonify({"error": True, "message": "不支持的操作系统，请选择 linux 或 windows", "code": 400}), 400

    confirm_ownership = _is_truthy(data.get("confirm_ownership"))
    confirm_execute = _is_truthy(data.get("confirm_execute"))

    # APPLY 双确认闸门（前端已拦，后端二次校验——R4 防线不可绕过）
    if mode == "apply" and (not confirm_ownership or not confirm_execute):
        return jsonify(
            {
                "error": True,
                "message": "[R4] APPLY 真机执行需要同时确认所有权和执行授权。"
                "请勾选「我确认拥有该资产」和「确认在真机执行加固」。",
                "code": 400,
            }
        ), 400

    # ---- 加载扫描记录 → 获取 target ----
    core = current_app.config["LIGHTSHIELD_CORE"]
    scan = core.load_scan(scan_id)
    if scan is None:
        return jsonify({"error": True, "message": f"扫描记录不存在: {scan_id}", "code": 404}), 404

    target = scan.target

    # ---- 执行闭环 ----
    session["harden_confirmed_at"] = datetime.now(timezone.utc).isoformat()

    logger.info("web", f"闭环启动: scan_id={scan_id} target={target} mode={mode}")

    try:
        result = core.run_harden_closed_loop(
            target=target,
            os_platform=os_platform,
            confirm_ownership=confirm_ownership,
            mode=mode,
            confirm_execute=confirm_execute,
        )
    except Exception as exc:
        logger.error("web", f"闭环执行异常: scan_id={scan_id} {exc}")
        return jsonify({"error": True, "message": f"闭环执行失败：{exc}", "code": 500}), 500

    # ---- 构造响应 ----
    response_data = result.to_dict()
    response_data["success"] = result.overall in ("verified", "partial", "generated_only")
    response_data["scan_id"] = scan_id

    status_code = 200 if result.overall != "failed" else 422  # 422 Unprocessable Entity → 前端展示失败但不崩

    logger.info(
        "web",
        f"闭环完成: scan_id={scan_id} overall={result.overall} "
        f"resolved={len(result.verification.get('resolved', [])) if result.verification else 0}",
    )
    return jsonify(response_data), status_code


@api_bp.route("/script/<scan_id>/<path:filename>", methods=["GET"])
@login_required
def api_download_script(scan_id: str, filename: str):
    """下载生成的加固 / 回滚脚本。"""
    logger = get_logger()

    if not _is_allowed_script_filename(filename):
        logger.warning("web", f"拒绝脚本下载：非法文件名 scan_id={scan_id} filename={filename}")
        return jsonify({"error": True, "message": "脚本文件名不在下载白名单内", "code": 400}), 400

    if not _validate_download_csrf():
        logger.warning("web", f"拒绝脚本下载：CSRF 校验失败 scan_id={scan_id} filename={filename}")
        return jsonify({"error": True, "message": "CSRF 校验失败，请刷新页面后重试", "code": 403}), 403

    if not session.get("harden_confirmed_at"):
        logger.warning("web", f"拒绝脚本下载：未确认 R4 所有权 scan_id={scan_id} filename={filename}")
        return jsonify({"error": True, "message": "[R4] 下载脚本前请先确认目标所有权或授权范围", "code": 403}), 403

    config: LightShieldConfig = current_app.config["LIGHTSHIELD_CONFIG"]
    script_path = _resolve_script_path(config.report_output_dir, filename)
    if script_path is None:
        logger.warning("web", f"脚本文件不存在：scan_id={scan_id} filename={filename}")
        return jsonify({"error": True, "message": "脚本文件不存在", "code": 404}), 404

    logger.info("web", f"脚本下载：scan_id={scan_id} filename={filename}")
    return send_from_directory(
        directory=script_path.parent,
        path=script_path.name,
        as_attachment=True,
        download_name=script_path.name,
        mimetype="application/octet-stream",
        max_age=0,
    )


# =============================================================================
# 辅助函数：从仓库数据重建领域对象
# =============================================================================


def _format_sse(data: dict, event: str | None = None) -> str:
    """Format one Server-Sent Event frame."""
    prefix = f"event: {event}\n" if event else ""
    return f"{prefix}data: {json.dumps(data, ensure_ascii=False)}\n\n"


def _is_allowed_script_filename(filename: str) -> bool:
    """Return True when filename matches the hardening script whitelist."""
    if not filename or "/" in filename or "\\" in filename:
        return False
    if Path(filename).name != filename:
        return False
    return any(fnmatchcase(filename, pattern) for pattern in SCRIPT_FILENAME_PATTERNS)


def _validate_download_csrf() -> bool:
    """Validate CSRF for GET-based script downloads."""
    expected = session.get("_csrf_token")
    supplied = (
        request.headers.get("X-CSRF-Token")
        or request.args.get("_csrf_token")
        or request.args.get("csrf_token")
        or request.form.get("_csrf_token")
    )
    return bool(expected and supplied and secrets.compare_digest(str(expected), str(supplied)))


def _resolve_script_path(output_dir: str, filename: str) -> Path | None:
    """Resolve a script path under report_output_dir without allowing traversal."""
    base_dir = Path(output_dir).expanduser().resolve(strict=False)
    script_path = (base_dir / filename).resolve(strict=False)
    if script_path.parent != base_dir or not script_path.is_file():
        return None
    return script_path


def _script_basename(path: str | None) -> str:
    """Return a generated script basename suitable for download URL construction."""
    if not path:
        return ""
    return Path(path).name


def _is_truthy(value) -> bool:
    """Return True for common JSON/form truthy values."""
    if isinstance(value, bool):
        return value
    if isinstance(value, str):
        return value.lower() in {"1", "true", "yes", "on"}
    return bool(value)